+++ samples/SampleClient/MyClient.cs
@@ -1,6 +1,7 @@
-﻿using Ether.Network;
-using System;
+﻿using Ether.Network.Client;
 using Ether.Network.Packets;
+using System;
+using System.Net.Sockets;
 
 namespace SampleClient
 {
@@ -12,20 +13,21 @@ namespace SampleClient
         /// <param name="host"></param>
         /// <param name="port"></param>
         /// <param name="bufferSize"></param>
-        public MyClient(string host, int port, int bufferSize) 
-            : base(host, port, bufferSize)
+        public MyClient(string host, int port, int bufferSize)
         {
+            this.Configuration.Host = host;
+            this.Configuration.Port = port;
+            this.Configuration.BufferSize = bufferSize;
         }
 
         /// <summary>
         /// Handles incoming messages.
         /// </summary>
         /// <param name="packet"></param>
-        protected override void HandleMessage(NetPacketBase packet)
+        public override void HandleMessage(INetPacketStream packet)
         {
             var response = packet.Read<string>();
-
-            Console.WriteLine("-> Server response: {0}", response);
+            Console.WriteLine($"-> Server response: '{response}'");
         }
 
         /// <summary>
@@ -33,7 +35,7 @@ namespace SampleClient
         /// </summary>
         protected override void OnConnected()
         {
-            Console.WriteLine("Connected to {0}", this.Socket.RemoteEndPoint.ToString());
+            Console.WriteLine("Connected to {0}", this.Socket.RemoteEndPoint);
         }
 
         /// <summary>
@@ -42,6 +44,15 @@ namespace SampleClient
         protected override void OnDisconnected()
         {
             Console.WriteLine("Disconnected");
+        }
+
+        /// <summary>
+        /// Triggered when an error occurs.
+        /// </summary>
+        /// <param name="socketError"></param>
+        protected override void OnSocketError(SocketError socketError)
+        {
+            Console.WriteLine("Socket Error: {0}", socketError.ToString());
         }
     }
 }