+++ samples/SampleServer/Client.cs
@@ -1,10 +1,10 @@
-﻿using Ether.Network;
-using System;
+﻿using Ether.Network.Common;
 using Ether.Network.Packets;
+using System;
 
 namespace SampleServer
 {
-    internal sealed class Client : NetConnection
+    internal sealed class Client : NetUser
     {
         /// <summary>
         /// Send hello to the incoming clients.
@@ -23,16 +23,16 @@ namespace SampleServer
         /// Receive messages from the client.
         /// </summary>
         /// <param name="packet"></param>
-        public override void HandleMessage(NetPacketBase packet)
+        public override void HandleMessage(INetPacketStream packet)
         {
-            string value = packet.Read<string>();
+            var value = packet.Read<string>();
 
             Console.WriteLine("Received '{1}' from {0}", this.Id, value);
 
             using (var p = new NetPacket())
             {
                 p.Write(string.Format("OK: '{0}'", value));
-                this.Send(p);
+                this.Server.SendToAll(p);
             }
         }
     }