+++ samples/SampleServer/SampleServer.cs
@@ -1,4 +1,4 @@
-﻿using Ether.Network;
+﻿using Ether.Network.Server;
 using System;
 
 namespace SampleServer
@@ -11,9 +11,11 @@ namespace SampleServer
         public SampleServer()
         {
             this.Configuration.Backlog = 100;
-            this.Configuration.Port = 8888;
+            this.Configuration.Port = 4444;
             this.Configuration.MaximumNumberOfConnections = 100;
             this.Configuration.Host = "127.0.0.1";
+            this.Configuration.BufferSize = 8;
+            this.Configuration.Blocking = true;
         }
 
         /// <summary>