+++ scripts/build.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-dotnet restore && dotnet build ./src/Ether.Network/Ether.Network.csproj -f netstandard1.3