+++ src/Ether.Network/Packets/NetPacket.cs
@@ -1,34 +1,19 @@
-﻿using Ether.Network.Exceptions;
-using System;
-using System.Collections.Generic;
-using System.IO;
+﻿using System.IO;
 
 namespace Ether.Network.Packets
 {
     /// <summary>
     /// Represents a Ehter.Network built-in packet.
     /// </summary>
-    public sealed class NetPacket : NetPacketBase
+    public sealed class NetPacket : NetPacketStream
     {
-        /// <summary>
-        /// Gets the packet buffer.
-        /// </summary>
-        public override byte[] Buffer
-        {
-            get
-            {
-                long oldPosition = this.Position;
+        private readonly int HeaderSize = sizeof(int);
 
-                this.MemoryWriter.Seek(0, SeekOrigin.Begin);
-                this.Write(this.Size - sizeof(int));
-                this.MemoryWriter.Seek((int)oldPosition, SeekOrigin.Begin);
-
-                return this.GetBuffer();
-            }
-        }
+        /// <inheritdoc />
+        public override byte[] Buffer => this.BuildBuffer();
 
         /// <summary>
-        /// Creates a new NetPacket in write-only mode.
+        /// Creates a new <see cref="NetPacket"/> in write-only mode.
         /// </summary>
         public NetPacket()
         {
@@ -36,7 +21,7 @@ namespace Ether.Network.Packets
         }
 
         /// <summary>
-        /// Creates a new NetPacket in read-only mode.
+        /// Creates a new <see cref="NetPacket"/> in read-only mode.
         /// </summary>
         /// <param name="buffer"></param>
         public NetPacket(byte[] buffer)
@@ -45,36 +30,18 @@ namespace Ether.Network.Packets
         }
 
         /// <summary>
-        /// Split the incoming packets.
+        /// Builds the final buffer.
         /// </summary>
-        /// <param name="buffer"></param>
         /// <returns></returns>
-        public static IReadOnlyCollection<NetPacketBase> Split(byte[] buffer)
+        private byte[] BuildBuffer()
         {
-            var packets = new List<NetPacket>();
+            long oldPosition = this.Position;
 
-            using (var memoryStream = new MemoryStream(buffer))
-            using (var readerStream = new BinaryReader(memoryStream))
-            {
-                try
-                {
-                    while (readerStream.BaseStream.Position < readerStream.BaseStream.Length)
-                    {
-                        var packetSize = readerStream.ReadInt32();
-                        
-                        if (packetSize == 0)
-                            break;
-                        
-                        packets.Add(new NetPacket(readerStream.ReadBytes(packetSize)));
-                    }
-                }
-                catch (Exception e)
-                {
-                    throw new EtherPacketException("An error occured while splitting the incoming packets.", e);
-                }
-            }
+            this.Seek(0, SeekOrigin.Begin);
+            this.Write(this.Size - HeaderSize);
+            this.Seek((int)oldPosition, SeekOrigin.Begin);
 
-            return packets;
+            return base.Buffer;
         }
     }
 }