+++ CHANGELOG.md
@@ -2,6 +2,42 @@
 
 ## [Unreleased]
 
+## [3.0.0] - 16/03/2018
+
+** Breaking Changes **
+
+### Added
+
+- New architecture
+- Add `INetPacketStream` interface and `NetPacketStream` implementation for read/write packets.
+- Add `NetUser` interface. Inherits from `NetConnection`.
+- Add `OnSocketError()` method to `NetClient`. (PR [#42](https://github.com/Eastrall/Ether.Network/pull/42))
+- Add all interfaces support to `NetServer`. (PR [#42](https://github.com/Eastrall/Ether.Network/pull/42))
+- Add support for reading byte arrays with `INetPacketStream`. (PR [#48](https://github.com/Eastrall/Ether.Network/pull/48))
+- Add broadcast system for `NetServer`. (PR [#52](https://github.com/Eastrall/Ether.Network/pull/52))
+- Add `NetClientConfiguration` (PR [#78](https://github.com/Eastrall/Ether.Network/pull/78))
+- Add `GetUser(Guid)` to `INetServer` (PR [#85](https://github.com/Eastrall/Ether.Network/pull/85))
+
+### Changed
+
+- `NetConnection` is lighter.
+- `NetClient` inherits from `NetUser`
+- `NetClient` sending process is using a queue.
+- `NetClient` message handler process is using a queue.
+- `NetServer` inherits from `NetConnection`.
+- Receive algorithm review.
+
+### Fix
+
+- Improve receive process on `NetClient` and `NetServer`.
+- Fix disconnect process on `NetClient`.
+
+### Removed
+
+- `NetPacketBase` has been removed and replaced with `NetPacketStream`.
+
+## [Released]
+
 ## [2.1.0-pre] - 2017-10-08
 
 ### Added
@@ -21,8 +57,6 @@
 ### Removed
 
 - Buffer Manager
-
-## [Released]
 
 ## [2.0.1] - 2017-08-06